Add support for adding green spaces layer
Support visualizing green spaces in the Helsinki capital region
alongside the price data. The green spaces layer can be toggled using a
checkbox in a new "Map Layers" dialog. The green space data is fetched
using a static bounding box for the capital region and is cached using a
static cache key at the moment.