Build and push container images in GH Actions
Extend the CI pipeline by adding build and push steps for the frontend &
backend images. Tagging is determined based on the version in
package.json.

Push images to GHCR instead of ACR. GHCR is free for public packages
while ACR is not. This also allows keeping more of the project in
public.